+++ src/main/java/com/tjx/lew00305/slimstore/model/report/TransactionTenderAggregationInterface.java
@@ -1,5 +1,7 @@
 package com.tjx.lew00305.slimstore.model.report;
 
+import java.math.BigDecimal;
+
 public interface TransactionTenderAggregationInterface {
 
     Integer getStore();
@@ -7,7 +9,7 @@ public interface TransactionTenderAggregationInterface {
     String getDate();
     Integer getReg();
     String getType();
-    Float getValue();
+    BigDecimal getValue();
   
 }
